% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03BaseClass.R
\name{addpreprocessor}
\alias{addpreprocessor}
\title{addpreprocessor}
\usage{
addpreprocessor(classname, operation, mode = "numeric")
}
\arguments{
\item{classname}{(character)}

\item{operation}{(expression as character string)}

\item{mode}{(character) default to "numeric" for operation to be computed on data frame with numeric variables, option="all" for DataClass object with slots x (numeeric variables) and y (factor of class labels)}
}
\description{
addpreprocessor is a helper function for inheriting subclasses from BaseClass
}
\examples{
## Set of examples using only numeric variables and no class labels
## addpreprocessor("naomit", "na.omit(basedata)")
## addpreprocessor("scale", "scale(basedata,center=FALSE)")
## addpreprocessor("nooutlierremove", "identity(basedata)")
## addpreprocessor("softmaxscale", "data.frame(apply(basedata, 2, DMwR::SoftMax))")
##
## An example using also class labels and a supporting function
## addpreprocessor("randomforestimpute", "rfimputefunc(basedata)", mode="all")
}

