% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/06AnalysisClass.R
\docType{class}
\name{AnalysisClass-class}
\alias{AnalysisClass-class}
\title{AnalysisClass}
\description{
AnalysisClass represents the analysis of preprocessing combinations
}
\section{Slots}{

\describe{
\item{\code{best}}{(data frame) best (that is, lowest misclassification error) combination}

\item{\code{all}}{(data frame) all preprocessing combinations and respective misclassification errors}

\item{\code{importance}}{(data frame) variable importance of phases in predicting misclassification error}

\item{\code{rules}}{assosiation rules for "low" (that is, last 10 percent) misclassification error}
}}

