% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/06AnalysisClass.R
\name{analyze}
\alias{analyze}
\title{analyse}
\usage{
analyze(out)
}
\arguments{
\item{out}{(data frame) out of function preprocomb or preproadeq}
}
\description{
analyze outputs the analysis of preprocessing combinations
}
\examples{
gridclassobject <- initializegridclassobject(list("outlier", "selection"), iris)
predictioncontrol <- initializepredictioncontrolclassobject(predictors='rf', gridclassobject)
out <- preprocomb(predictioncontrol)
analysis <- analyze(out)
analysis@best
}

