% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03BaseClass.R
\docType{class}
\name{BaseClass-class}
\alias{BaseClass-class}
\title{BaseClass}
\description{
BaseClass represents the preprocessors
}
\section{Slots}{

\describe{
\item{\code{objectname}}{(character) name of the object}

\item{\code{objectoperation}}{(expression as character string) expression to be computed}

\item{\code{isvalid}}{(logical) TRUE, if computation of objectoperation is successful}

\item{\code{data}}{(DataClass) data to be used in object operation computation}
}}

