% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01DataClass.R
\docType{class}
\name{DataClass-class}
\alias{DataClass-class}
\title{DataClass}
\description{
DataClass the data
}
\section{Slots}{

\describe{
\item{\code{x}}{(data frame)}

\item{\code{y}}{(factor) class labels}
}}

