% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00Utils.R
\docType{data}
\name{examplecombgrid}
\alias{examplecombgrid}
\title{setgrid example}
\format{GridClass}
\usage{
examplecombgrid
}
\description{
Contaminated Iris-data preprocessed with 90 combinations. \cr
examplecombgrid <- setgrid(phases=c("imputation", "scaling", "smoothing"), data=contaminateddf) \cr
}
\keyword{datasets}

