% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00Utils.R
\docType{data}
\name{examplemetaheur}
\alias{examplemetaheur}
\title{metaheur example}
\format{A MetaheurClass object}
\usage{
examplemetaheur
}
\description{
examplemetaheur <- metaheur(examplegrid, model="svmRadial", iterations = 30, nholdout = 400) \cr
}
\keyword{internal}

