% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00Utils.R
\docType{data}
\name{examplesimulation}
\alias{examplesimulation}
\title{preprosim example}
\format{A PreprosimClass object}
\usage{
examplesimulation
}
\description{
examplesimulation <- preprosimrun(iris, fitmodels=FALSE) # 6561 contaminated data sets \cr
}
\keyword{internal}

