% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04GridClass.R
\name{formdata}
\alias{formdata}
\title{formdata}
\usage{
formdata(grid, data)
}
\arguments{
\item{grid}{(GridClass)}

\item{data}{(DataClass)}
}
\description{
formdata function takes grid and a DataClass object as arguments and computes the data for each row in grid
}

