% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04GridClass.R
\name{getcombinations}
\alias{getcombinations}
\title{get preprocessing combinations from a grid}
\usage{
getcombinations(gridobject)
}
\arguments{
\item{gridobject}{(grid class) object}
}
\description{
get preprocessing combinations from a grid
}
\examples{
head(getcombinations(examplecombgrid))
}

