% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03PreprocessorClass.R
\name{getpreprocessor}
\alias{getpreprocessor}
\title{gets preprocessor definition}
\usage{
getpreprocessor(preprocessor = NULL)
}
\arguments{
\item{preprocessor}{(character) name of preprocessor, defaults to NULL for list of all preprocessors}
}
\description{
gets preprocessor definition
}
\details{
getpreprocessor with the name of the preprocessor shows preprocessing function body defined with setpreprocessor().\cr
getpreprocessor without name shows all preprocessors that can be used by functions prepro() and setphase().
}
\examples{
getpreprocessor()
getpreprocessor("basicscale")
}

