% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03PreprocessorClass.R
\name{getpreprocessors}
\alias{getpreprocessors}
\title{get available preprocessors and their descriptions}
\usage{
getpreprocessors(type = "description", nro)
}
\arguments{
\item{type}{(character) default "description", alternatives "name" or "definition"}

\item{nro}{(integer) combination number}
}
\description{
get available preprocessors and their descriptions
}

