% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04GridClass.R
\name{getpreprocombdf}
\alias{getpreprocombdf}
\title{get preprocessed data set from a grid}
\usage{
getpreprocombdf(gridobject, nro, type = "dataframe")
}
\arguments{
\item{gridobject}{(grid class object)}

\item{nro}{(integer) combination number}

\item{type}{(character) default "dataframe", option "summary"}
}
\description{
get preprocessed data set from a grid
}
\examples{
str(getpreprocombdf(examplecombgrid, 3))
}

