% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05PredictionControlClass.R
\name{getprogrammaticprediction}
\alias{getprogrammaticprediction}
\title{getprogrammaticprediction output classification accuracy. This function is exported
to be used by package metaheur.}
\usage{
getprogrammaticprediction(preprocesseddataset, predictors, nholdout)
}
\arguments{
\item{preprocesseddataset}{(DataClass)}

\item{predictors}{caret models}

\item{nholdout}{number of holdout rounds}
}
\description{
getprogrammaticprediction output classification accuracy. This function is exported
to be used by package metaheur.
}

