% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04GridClass.R
\docType{class}
\name{GridClass-class}
\alias{GridClass-class}
\title{GridClass}
\description{
GridClass represents the grid
}
\section{Slots}{

\describe{
\item{\code{grid}}{(data frame) combinations of preprocessors}

\item{\code{data}}{(list) list of data frames computed from each row in grid}
}}

