% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01DataClass.R
\name{initializedataclassobject}
\alias{initializedataclassobject}
\title{constructor function for creating a DataClass object}
\usage{
initializedataclassobject(data)
}
\arguments{
\item{data}{(data.frame)}
}
\description{
initializedataclassobject() is a constructor function for creating a DataClass object. The main
use case is adding of new preprocessing techniques to the framework by the user. Added preprocessing
techniques (i.e. functions) take as input and must return a DataClass object. See setpreprocessor().
}
\details{
Argument 'data' must have only numeric columns and one factor column.
}
\examples{
## dataobject <- initializedataclassobject(iris)
}

