% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01DataClass.R
\name{initializedataobject}
\alias{initializedataobject}
\title{initializedataobject}
\usage{
initializedataobject(data)
}
\arguments{
\item{data}{(data frame)}
}
\value{
(DataClass) object
}
\description{
initializedataobject is a constructor function for initializing a DataClass object.¨¨
The function separates class labels from rest of the (numerical) variables.
}

