% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04GridClass.R
\name{initializegridclassobject}
\alias{initializegridclassobject}
\title{initializegridclassobject}
\usage{
initializegridclassobject(phases, data)
}
\arguments{
\item{phases}{(character) vector}

\item{data}{(data frame)}
}
\description{
initializegridclassobject is a constructor function for initializing a GridClass object.
}
\examples{
gridclassobject <- initializegridclassobject(phases=c("outlier", "selection"), data=iris)
}

