% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02PhaseClass.R
\name{initializephaseclassobject}
\alias{initializephaseclassobject}
\title{initializephaseclassobject}
\usage{
initializephaseclassobject(phasename, preprocessor, preimpute)
}
\arguments{
\item{phasename}{(character)}

\item{preprocessor}{(list) subclass objects}

\item{preimpute}{(logical) whether phase is executed before imputation}
}
\description{
initializephaseclassobject is a constructor function for initializing a PhaseClass object.
The function separates class labels from rest of the variables.
}

