% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05PredictionControl.R
\name{initializepredictioncontrolclassobject}
\alias{initializepredictioncontrolclassobject}
\title{initializepredictioncontrolclassobject}
\usage{
initializepredictioncontrolclassobject(predictors, grid)
}
\arguments{
\item{predictors}{(character) vector of predictors}

\item{grid}{(GridClass)}
}
\description{
initializepredictioncontrolclassobject is a constructor function for initializing a PredictionControlClass object.
}

