% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04GridClass.R
\name{initializesubclassobject}
\alias{initializesubclassobject}
\title{initializesubclassobject}
\usage{
initializesubclassobject(classname, dataobject)
}
\arguments{
\item{classname}{(character)}

\item{dataobject}{(DataClass or sub class of BaseClass)}
}
\description{
initializesubclassobject is a constructor function for initializing sub class objects from BaseClass.
}

