% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02PhaseClass.R
\docType{class}
\name{PhaseClass-class}
\alias{PhaseClass-class}
\title{PhaseClass}
\description{
PhaseClass represents the phases of preprocessing such as missing value imputation, outlier removal etc.
}
\section{Slots}{

\describe{
\item{\code{objectname}}{(character) name of the phase}

\item{\code{preprotransformations}}{(list) preprocessing transformations belonging to the phase}

\item{\code{preimpute}}{(logical) TRUE, if phase is execution before missing value imputation}
}}

