% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05PredictionControl.R
\docType{class}
\name{PredictionClass-class}
\alias{PredictionClass-class}
\title{PredictionClass}
\description{
PredictionClass stores the computation of preprocomb() or preproadeq()
}
\section{Slots}{

\describe{
\item{\code{predictors}}{(character) vector of predictors}
}}

