% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05PredictionControl.R
\docType{class}
\name{PredictionControl-class}
\alias{PredictionControl-class}
\title{PredictionControlClass}
\description{
PredictionControlClass controls as argument the computations
}
\section{Slots}{

\describe{
\item{\code{predictors}}{(character) vector of predictors}

\item{\code{grid}}{(GridClass)}
}}

