% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03PreprocessorClass.R
\name{prepro}
\alias{prepro}
\title{prepro}
\usage{
prepro(dataobject, classname, model = "knn", nholdout = 2,
  nsharehopkins = 3)
}
\arguments{
\item{dataobject}{(sub class/ data frame/ DataClass) object}

\item{classname}{(character) name of preprocessor (i.e. PreprocessorClass sub class as defined by setpreprocessor())}

\item{model}{(character) caret model name, note: the required model library must be attached, defaults to "knn"}

\item{nholdout}{(integer) number of holdout rounds used in computation of classification accuracy, must be two or more, defaults to two}

\item{nsharehopkins}{(integer) denominator for sample size for hopkins statistics, defauls to three  (n=nrow(data)/3)}
}
\value{
object of PreprocessorClass sub class
}
\description{
prepro is the main function for interactive use. It takes data, transforms it according to the given
preprocessor and computes statistics of the transformed data. The main use case is the chaining of
the preprocessors as show in the examples below.
}
\examples{
## a <- prepro(iris, "basicscale")
## b <- prepro(a, "rfselect75")
## d <- prepro(iris, "scale", "rf", 20, 2, 10)
}

