% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05PredictionControl.R
\name{preproadeq}
\alias{preproadeq}
\title{preproadeq}
\usage{
preproadeq(predictioncontrol)
}
\arguments{
\item{predictioncontrol}{(PredictionControlClass)}
}
\description{
preproadeq is the supplementary execution function for computing the adequacy of the (that is, learnability) of each row in grid
}

