% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00Utils.R
\docType{data}
\name{preprocessordefinitionstorage}
\alias{preprocessordefinitionstorage}
\title{environment for storing preprocessor definitions}
\format{An object of class \code{environment} of length 19.}
\usage{
preprocessordefinitionstorage
}
\description{
an environment to get the preprocessing technique function bodies
}
\keyword{datasets}

