% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07AnalysisComponent.R
\name{preprocombplot}
\alias{preprocombplot}
\title{plot preprocessing combination evaluation results}
\usage{
preprocombplot(ResultClassobject, type = "boxplot")
}
\arguments{
\item{ResultClassobject}{(ResultClass) object to be plotted}

\item{type}{(character) boxplot or lineplot}
}
\description{
plot preprocessing combination evaluation results
}

