% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08DefaultPreprocessorsAndPhases.R
\docType{data}
\name{preprodefault}
\alias{preprodefault}
\title{six default phases with preprocessing techniques}
\format{An object of class \code{character} of length 6.}
\usage{
preprodefault
}
\description{
Totals 1080 combinations. preprodefault object can be used as default phases for setgrid().
}
\examples{
## grid <- setgrid(preprodefault, iris)
}
\keyword{internal}

