% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08DefaultPreprocessorsAndPhases.R
\docType{data}
\name{preprodefaultextended}
\alias{preprodefaultextended}
\title{eigth default phases with preprocessing techniques}
\format{An object of class \code{character} of length 7.}
\usage{
preprodefaultextended
}
\description{
Totals 6480 combinations. preprodefaultextended object can be used as default phases for setgrid().
}
\examples{
## grid <- setgrid(preprodefaultextended, iris[1:90,])
}
\keyword{internal}

