% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/07AnalysisClass.R
\name{preprorules}
\alias{preprorules}
\title{preprorules}
\usage{
preprorules(preprocombobject, support = 0.05, confidence = 0.7)
}
\arguments{
\item{preprocombobject}{(PreProCombClass)}

\item{support}{(numeric) defaults to 0.05}

\item{confidence}{(numeric) defaults to 0.7}
}
\description{
preprorules shows the association rules for best ten percent misclassification errors
preprorules uses arules package
}

