% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/07AnalysisClass.R
\name{preprotree}
\alias{preprotree}
\title{preprotree}
\usage{
preprotree(preprocombobject)
}
\arguments{
\item{preprocombobject}{(PreProCombClass)}
}
\description{
preprotree shows c5.0 tree for best ten percent misclassification errors
preprorules uses C50 package
}

