% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04GridClass.R
\name{setgrid}
\alias{setgrid}
\title{setgrid}
\usage{
setgrid(phases, data)
}
\arguments{
\item{phases}{(character) vector of phases}

\item{data}{(data frame)}
}
\description{
setgrid creates the combinations as a grid and computes the transformed data for each combination (row of grid)
}
\examples{
## grid <- setgrid(phases=c("outlier", "selection"), data=iris)
}

