% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02PhaseClass.R
\name{setphase}
\alias{setphase}
\title{setphase}
\usage{
setphase(phasename, preprocessor, preimpute)
}
\arguments{
\item{phasename}{(character) name of the phase}

\item{preprocessor}{(character) vector of preprocessors (see ?setpreprocessor) belonging to the phase}

\item{preimpute}{(logical) whether phase is executed before missing value imputation}
}
\value{
a PhaseClass object
}
\description{
setphase is a constructor function for defining a phase.
setphases initializes a PhaseClass object.
}
\details{
All elements of argument 'preprocessor' must point to PreprocessorClass objects constructed with function 'setpreprocessor()'.
}
\examples{
## imputation <- setphase("imputation", c("naomit", "meanimpute"), TRUE)
}

