% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03BaseClass.R
\name{setpreprocessor}
\alias{setpreprocessor}
\title{setpreprocessor}
\usage{
setpreprocessor(classname, operation, mode = "numeric")
}
\arguments{
\item{classname}{(character)}

\item{operation}{(expression as character string)}

\item{mode}{(character) default to "numeric" for operation to be computed on data frame with numeric variables, option="all" for DataClass object with slots x (numeeric variables) and y (factor of class labels)}
}
\description{
setpreprocessor is a helper function used to define sub classes of BaseClass.
Specifically, sub classes include the operation to be executed to preprocess data.
}
\details{
If operation deletes rows in numeric data such as outliers, the corresponding class
labels are deleted automatically.

If operation uses both numeric columns and class labels, the defined operation must
return both.
}
\examples{
## Set of examples using only numeric variables and no class labels
## setpreprocessor("naomit", "na.omit(basedata)")
## setpreprocessor("scale", "scale(basedata,center=FALSE)")
## setpreprocessor("nooutlierremove", "identity(basedata)")
## setpreprocessor("softmaxscale", "data.frame(apply(basedata, 2, DMwR::SoftMax))")
##
## An example using also class labels and a supporting function
## setpreprocessor("randomforestimpute", "rfimputefunc(basedata)", mode="all")
}

