% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/07AnalysisComponent.R
\name{showrules}
\alias{showrules}
\title{showrules}
\usage{
showrules(PreProCombClassobject, support = 0.05, confidence = 0.5)
}
\arguments{
\item{PreProCombClassobject}{(PreProCombClass)}

\item{support}{(numeric) support for association rules, default to 0.05}

\item{confidence}{(numeric) confidence for association rules, defaults to 0.5}
}
\description{
showrules shows association rules for classification accuracy.
Classification accuracy is categorical as 'high' for best twenty
percent and 'low' for rest.
}

