% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03PreprocessorClass.R
\name{testpreprocessors}
\alias{testpreprocessors}
\title{test preprocessing techniques against data}
\usage{
testpreprocessors(preprocessors = NULL, data = NULL)
}
\arguments{
\item{preprocessors}{(character) vector of preprocessors, by default gets all preprocessors with getpreprocessors()}

\item{data}{(data frame) to be tested against, defaults to random data frame without missing values}
}
\value{
boolean, TRUE is all data object validations are TRUE
}
\description{
Intended to be used when adding new preprocessing techniques with setpreprocessor().
}
\details{
testpreprocessor() with Iris as data argument is called from setpreprocessor()
}
\examples{
testpreprocessors()
}
\keyword{internal}

