% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03BaseClass.R
\name{transformdata}
\alias{transformdata}
\title{transformdata}
\usage{
transformdata(object, dataobject)
}
\arguments{
\item{object}{}

\item{dataobject}{}
}
\description{
transformdata is a generic function. Its methods are defined by setpreprocessor().
The function intented for package internal use.
}

