% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00Utils.R
\docType{data}
\name{exampleresult}
\alias{exampleresult}
\title{preprocomb example}
\format{ResultClass}
\usage{
exampleresult
}
\description{
Modified Iris-data preprocessed with 90 combinations and evaluated with \cr
svmRadial classifier and 400 times repreated holdout validation. \cr
}
\keyword{datasets}

