% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00Utils.R
\docType{data}
\name{exampleresult}
\alias{exampleresult}
\title{preprocomb example}
\format{A PreProCombClass object}
\usage{
exampleresult
}
\description{
An example grid object made of modified Iris-data with 3200 combinations \cr
evaluated with knn classifier classification accuracy and four holdrounds for each.
}
\keyword{datasets}

