% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03PreprocessorClass.R
\name{getpreprocessor}
\alias{getpreprocessor}
\title{gets preprocessor definition}
\usage{
getpreprocessor(preprocessor = NULL)
}
\arguments{
\item{preprocessor}{(character) name of preprocessor, defaults to NULL for list of all preprocessors
get preprocessor technique function body defined with setpreprocessor().
Shown preprocessors can be used by functions prepro() and setphase().}
}
\description{
gets preprocessor definition
}
\examples{
getpreprocessor()
getpreprocessor("basicscale")
}

