% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03PreprocessorClass.R
\name{getpreprocessors}
\alias{getpreprocessors}
\title{getpreprocessors}
\usage{
getpreprocessors()
}
\description{
gets the available preprocessors, that is: PreprocessorClass (sub) classes.
Shown preprocessors can be used by functions prepro() and setphase().
}

