% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05PredictionControlClass.R
\name{getprogrammaticprediction}
\alias{getprogrammaticprediction}
\title{parallel computation of classification accuracy holdout rounds}
\usage{
getprogrammaticprediction(preprocesseddataset, predictors, nholdout)
}
\arguments{
\item{preprocesseddataset}{(DataClass)}

\item{predictors}{caret models}

\item{nholdout}{number of holdout rounds}
}
\description{
This function is used internally and exported for package 'metaheur'.
}
\details{
If model tuning fails, NA is returned as classification accuracy of a combination.
If model fitting and prediction for holdout round fails, NA is returned for the holdout round.
}
\keyword{internal}

