% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04GridClass.R
\docType{class}
\name{GridClass-class}
\alias{GridClass-class}
\title{An S4 class representing (unevaluated) preprocessing combinations}
\description{
Preprocessing techniques defined with setpreprocessor() can be combined to a phase.
Phases defined with setphase() can be combined to a grid of combinations with setgrid().
The main programmatic use with preprocomb() takes a GridClass object as argument.
}
\details{
GridClass is also an interface for extending the system to package 'metaheur', which takes
a GridClass object to find near-optiomal combinations fast.
}
\section{Slots}{

\describe{
\item{\code{grid}}{(data frame) preprocessor combinations}

\item{\code{data}}{(list) DataClass objects}

\item{\code{validation}}{(data frame) validation results}
}}

