% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04GridClass.R
\docType{class}
\name{GridClass-class}
\alias{GridClass-class}
\title{GridClass}
\description{
GridClass is a container for preprocessor combinations and the corresponding preprocessed data sets.
GridClass is an interface for extending the system.
}
\details{
Extensions can include approximate combinatorial optimization for finding near-best
combinations faster.
}
\section{Slots}{

\describe{
\item{\code{grid}}{(data frame) preprocessor combinations}

\item{\code{data}}{(list) DataClass objects}
}}

