% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01DataClass.R
\name{initializedataclassobject}
\alias{initializedataclassobject}
\title{initializedataclassobject}
\usage{
initializedataclassobject(data)
}
\arguments{
\item{data}{(data.frame)}
}
\description{
initializedataclassobject is a function to construct a DataClass object.
The main use case is to return a DataClass object from user-defined preprocessors
in setpreprocessor()
}
\details{
Argument 'data' must have only numeric columns and one factor column.
}
\examples{
## dataobject <- initializedataclassobject(data.frame(x=scale(dataobject@x), dataobject@y))
}

