% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01DataClass.R
\name{initializedataclassobject}
\alias{initializedataclassobject}
\title{initialize a DataClass object}
\usage{
initializedataclassobject(data)
}
\arguments{
\item{data}{(data.frame)}
}
\description{
initializedataclassobject is used when new preprocessing techniques are added to the
system by the user. Added preprocessing techniques (i.e. functions) must return a
DataClass object.
}
\details{
Argument 'data' must have only numeric columns and one factor column.
}
\examples{
## dataobject <- initializedataclassobject(iris)
}

