% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03PreprocessorClass.R
\docType{class}
\name{PreprocessorClass-class}
\alias{PreprocessorClass-class}
\title{A virtual S4 class representing preprocessors}
\description{
Inheritance is controlled by setpreprocessor() function.
}
\section{Slots}{

\describe{
\item{\code{objectname}}{(character) object name}

\item{\code{objectoperation}}{(character) operation (R expression as character string)}

\item{\code{data}}{(DataClass) object}

\item{\code{classificationaccuracy}}{(numeric) classification accuracy}

\item{\code{hopkinsstatistic}}{(numeric) clustering tendency}

\item{\code{callhistory}}{(character) vector of current and previous calls}
}}

