% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03BaseClass.R
\docType{class}
\name{PreprocessorClass-class}
\alias{PreprocessorClass-class}
\title{PreprocessorClass}
\description{
PreprocessorClass is a virtual class. It is intented for package internal use.
}
\section{Slots}{

\describe{
\item{\code{objectname}}{(character) object name}

\item{\code{objectoperation}}{(character) operation (expression as character string)}

\item{\code{data}}{(DataClass) object}

\item{\code{classificationaccuracy}}{(numeric) classification accuracy}

\item{\code{hopkinsstatistic}}{(numeric) clustering tendency}

\item{\code{LOFskewness}}{(numeric) skewness value of LOF scores}

\item{\code{callhistory}}{(character) vector of current and previous calls}
}}

