% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03PreprocessorClass.R
\docType{class}
\name{PreprocessorClass-class}
\alias{PreprocessorClass-class}
\title{PreprocessorClass}
\description{
PreprocessorClass is an abstract class from which concrete preprocessor (sub) classes are inhereted.
Inheritance is controlled by setpreprocessor() function.
}
\section{Slots}{

\describe{
\item{\code{objectname}}{(character) object name}

\item{\code{objectoperation}}{(character) operation (expression as character string)}

\item{\code{data}}{(DataClass) object}

\item{\code{classificationaccuracy}}{(numeric) classification accuracy}

\item{\code{hopkinsstatistic}}{(numeric) clustering tendency}

\item{\code{ORHskewness}}{(numeric) skewness value of LOF scores}

\item{\code{callhistory}}{(character) vector of current and previous calls}
}}

