% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/06PreProCombClass.R
\name{preprocomb}
\alias{preprocomb}
\title{preprocomb}
\usage{
preprocomb(predictors, grid, nholdout = 1, search = "exhaustive")
}
\arguments{
\item{predictors}{(character) vector of predictors (names of models as defined in package caret)}

\item{grid}{(GridClass) object}

\item{nholdout}{(integer) number of holdout rounds, defaults to 1}

\item{search}{(character) defaults to "exhaustive" full blind search, "random" search 20 percent of grid, "grid" grid search 10 percent}
}
\description{
preprocomb outputs the analysis of preprocessing combinations
}
\examples{
## modifiediris <- droplevels(iris[-c(1:60),])
## grid <- setgrid(phases=c("outlier", "scaling", "sampling"), data=modifiediris)
## result <- preprocomb(predictors=c("svmRadial"), grid=grid, nholdout=2, search="grid")
## result@all
## result@rules
}

