% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/06AnalysisClass.R
\name{preprocomb}
\alias{preprocomb}
\title{preprocomb}
\usage{
preprocomb(predictors, grid, nholdout = 1)
}
\arguments{
\item{predictors}{(character) vector of predictors (names of models as defined in package caret)}

\item{grid}{(GridClass) object}

\item{nholdout}{(integer) number of holdout rounds}
}
\description{
preprocomb outputs the analysis of preprocessing combinations
}
\examples{
## grid <- setgrid(phases=c("outlier", "selection"), data=iris)
## result <- preprocomb(predictors=c('rf'), grid=grid, nholdout=1)
## result@best
## result@all
## result@tree
}

