% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05PredictionControl.R
\name{preprocomb}
\alias{preprocomb}
\title{preprocomb}
\usage{
preprocomb(predictioncontrol)
}
\arguments{
\item{predictioncontrol}{(PredictionControlClass)}
}
\description{
preprocomb is the main execution function for computing the misclassification rate of each preprocessed grid row
}

