% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/06AnalysisClass.R
\name{preprocomb}
\alias{preprocomb}
\title{preprocomb}
\usage{
preprocomb(predictioncontrol)
}
\arguments{
\item{predictioncontrol}{(PredictionControlClass)}
}
\description{
preprocomb outputs the analysis of preprocessing combinations
}
\examples{
gridclassobject <- initializegridclassobject(list("outlier", "selection"), iris)
predictioncontrol <- initializepredictioncontrolclassobject(predictors='rf', gridclassobject)
result <- preprocomb(predictioncontrol)

result@best
result@all
result@tree
}

