% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/06PreProCombClass.R
\name{preprocomb}
\alias{preprocomb}
\title{preprocomb}
\usage{
preprocomb(predictors, grid, nholdout = 2, search = "exhaustive")
}
\arguments{
\item{predictors}{(character) vector of predictors (names of models as defined in package caret), there must be an odd number of models}

\item{grid}{(GridClass) object representing the grid of combinations}

\item{nholdout}{(integer) number of holdout rounds, must be two or more, defaults to 2}

\item{search}{(character) defaults to "exhaustive" full blind search, "random" search 20 percent of grid, "grid" grid search 10 percent}
}
\description{
preprocomb the function of programmatic mode. It executes the computation of
misclassification errors, hopkins statistic and ORH outlier scores. The result
PreProCombClass object is used either to access its slots or to plot it with
preproplot()
}
\examples{
## modifiediris <- droplevels(iris[-c(1:60),])
## grid <- setgrid(phases=c("outlier", "scaling"), data=modifiediris)
## result <- preprocomb(predictors=c("svmRadial"), grid=grid, nholdout=2, search="exhaustive")
## result@allce
## result@allclustering
## result@alloutliers
##
## newphases <- c("outlier", "smoothing", "scaling", "selection", "sampling")
## newpredictors <- c("knn", "rf", "svmRadial")
## grid1 <- setgrid(phases=newphases, data=modifiediris)
## result1 <- preprocomb(predictors=newpredictors, grid=grid1, nholdout=2, search="grid")
}

