% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/06PreProCombClass.R
\docType{class}
\name{PreProCombClass-class}
\alias{PreProCombClass-class}
\title{PreProCombClass}
\description{
PreproCombClass is a container for output of executing preprocomb function.
It is an interface for extension of the system.
}
\section{Slots}{

\describe{
\item{\code{rawall}}{(data frame) all results}

\item{\code{catclassification}}{(data frame) classification accuracy categorized, "high" is more than 80 percent quantile value}

\item{\code{allclassification}}{(data frame) classification accuracy, means and standard deviations}

\item{\code{bestclassification}}{(data frame) best classification accuracy combinations}

\item{\code{allclustering}}{(data frame) hopkins statistics values}

\item{\code{bestclustering}}{(data frame) best hopkins statistics combinations}

\item{\code{alloutliers}}{(data frame) ORH outlier score for 95 percent quantile value}
}}

