% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/08DefaultPreprocessorsAndPhases.R
\docType{data}
\name{preprodefault}
\alias{preprodefault}
\title{preprodefault}
\format{\preformatted{ chr [1:7] "imputation" "variance" "smoothing" "scaling" "outliers" "sampling" ...
}}
\usage{
preprodefault
}
\description{
Default phases
}
\details{
preprodefault object can be used as default phases for setgrid
}
\keyword{datasets}

