% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04GridClass.R
\name{setgrid}
\alias{setgrid}
\title{setgrid}
\usage{
setgrid(phases, data)
}
\arguments{
\item{phases}{(character) vector of phases}

\item{data}{(data frame)}
}
\description{
setgrid takes the preprocessing phases, preprocessors and creates the combinations of them as a grid.
It then computes and stores the transformed data set for each combination.
setgrid initialized a GridClass object.
}
\examples{
## grid <- setgrid(phases=c("outlier", "selection"), data=iris)
}

