% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02PhaseClass.R
\name{setphase}
\alias{setphase}
\title{setphase}
\usage{
setphase(phasename, preprocessor, preimpute)
}
\arguments{
\item{phasename}{(character) name of the phase}

\item{preprocessor}{(character) vector of preprocessors (see ?setpreprocessor) belonging to the phase}

\item{preimpute}{(logical) whether phase is executed before missing value imputation}
}
\description{
setphase is a constructor function for initializing a PhaseClass object.
}
\examples{
## imputation <- setphase("imputation", c("naomit", "meanimpute"), TRUE)
}

