% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02PhaseClass.R
\name{setphase}
\alias{setphase}
\title{constructor function for defining a preprocessing phase.}
\usage{
setphase(phasename, preprocessor, preimpute)
}
\arguments{
\item{phasename}{(character) name of the phase}

\item{preprocessor}{(character) vector of preprocessors (see ?setpreprocessor) belonging to the phase}

\item{preimpute}{(logical) whether phase is missing value imputation}
}
\value{
a PhaseClass object
}
\description{
Preprocessing phases consist of preprocessing techniques defined with setpreprocessor().
Phases can be defined with setphase() and combined to a grid of combinations with setgrid().
}
\details{
All elements of argument 'preprocessor' must point to PreprocessorClass objects constructed with function 'setpreprocessor()'. \cr
If dataset contains missing values, missing value imputation must be the first phase.
}
\examples{
## imputation <- setphase("imputation", c("naomit", "meanimpute"), TRUE)
}

