% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03PreprocessorClass.R
\name{setpreprocessor}
\alias{setpreprocessor}
\title{constructor function for adding a new preprocessing technique to the system}
\usage{
setpreprocessor(classname, operation, description)
}
\arguments{
\item{classname}{(character) name of the preprocessing techniques}

\item{operation}{(expression as character string) expression to be computed on data}

\item{description}{(character)}
}
\value{
NULL, side-effect: definition of S4 class derived from PreprocessorClass and corresponding transformdata-method
}
\description{
The main argument is the operation that is executed to transform the data such as "na.omit(basedata)"
for removing rows that have missing values. An operation can process either only the numeric
columns or also the class label column.
}
\details{
Preprocessing techniques defined with setpreprocessor() can be combined to a phase.
Phases defined with setphase() can be combined to a grid of combinations with setgrid().

The user-defined S4 class definitions are stored in global environment and thus the function can not be used from an other package.

scaleexample <- function(dataobject) dataobject <- initializedataclassobject(data.frame(x=scale(dataobject@x), dataobject@y))
setpreprocessor("scaleexample", "scaleexample(dataobject)")

setpreprocessor() runs a test on test data frame and prints the outcome of the test.
}

