% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03BaseClass.R
\name{testpreprocessors}
\alias{testpreprocessors}
\title{testpreprocessors}
\usage{
testpreprocessors(preprocessors = NULL, data = NULL)
}
\arguments{
\item{preprocessors}{(character) vector of preprocessors, by default gets all preprocessors with getpreprocessors()}

\item{data}{(data frame) to be tested against, defaults to random data frame without missing values}
}
\description{
run a test for preprocessors
}

