% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03PreprocessorClass.R
\name{transformdata}
\alias{transformdata}
\title{transformdata}
\usage{
transformdata(object, dataobject)
}
\arguments{
\item{object}{(PreprocessorClass) object}

\item{dataobject}{(DataClass/data frame) object}
}
\description{
transformdata is a generic preprocessing function. Its methods are defined by setpreprocessor().
The function is intented for package internal use, but exported so that classes can be inhereted from it.
}

