% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03PreprocessorClass.R
\name{transformdata}
\alias{transformdata}
\title{transformdata}
\usage{
transformdata(object, dataobject)
}
\arguments{
\item{object}{(PreprocessorClass) object}

\item{dataobject}{(DataClass/data frame) object}
}
\description{
transformdata is a generic function. Its methods are defined by setpreprocessor().
The function is intented for package internal use.
}

